added some authentication for the users

+++ app/Http/Controllers/PostsController.php
@@ -6,6 +6,10 @@ use Illuminate\Http\Request;
 use App\Post;
 class PostsController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -39,10 +43,12 @@ class PostsController extends Controller
             'myTitle' => 'required',
             'body' => 'required'
         ]);
+        $id = auth()->user()->id;
 
             $post = new Post;
             $post->title = $request->input('myTitle');
             $post->body = $request->input('body');
+            $post->user_id = $id;
             $post->save();
 
             return redirect( route('posts.index') );
@@ -69,6 +75,9 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect( route('posts.index'));
+        }
         return view('posts.edit')->with('post', $post);   
     }
 
@@ -103,6 +112,9 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post =Post::find($id);
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect( route('posts.index'));
+        }
         $post->delete();
         return redirect( route('posts.index'));
     }
